+++ CHANGELOG.md
@@ -13,16 +13,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 * Exchange button and modal
 * Shapeshift endpoints
 * Account Type component for displaying active wallet logo
+* Link to Ledger support page on Ledger connection page
+* Ledger affiliate link to wallet selection screen
+* Donate ETH button and modal
+* Input fetching animation
 
 ### Changed
 
 * Fix txn fee conversion
 * Layout padding definition locations
+* Fix mobile layout on the wallet selection screen (#232)
 * DrodpownAsset improved logic
 * Refactor Send and Exchange modals/reducers for consistency
 * Fix unit time references
 * Moved missing language strings
-* Added fetching animation to input
+* Fix overflow padding for dropdowns
 
 ### Removed